+++ apps/hash-frontend/src/pages/api/block-generator/initialize-container.api.ts
