+++ apps/hash-frontend/src/pages/block-generator/bouncing-dots-loader.tsx
