+++ apps/hash-frontend/src/pages/block-generator/defaults.ts
