+++ apps/hash-frontend/src/pages/block-generator/demo-live-editor.tsx
