+++ apps/hash-frontend/src/pages/block-generator/preview/App.js
@@ -10,5 +10,3 @@ const MyButton = () => {
 };
 
 export default MyButton;
-
-// Dependencies: ['@mui/material']
