+++ apps/hash-frontend/src/pages/block-generator/preview/package.json
