+++ apps/hash-frontend/src/pages/block-generator/preview/public/favicon.ico
