+++ apps/hash-frontend/src/pages/block-generator/preview/public/index.html
