+++ apps/hash-frontend/src/pages/block-generator/preview/public/logo192.png
