+++ apps/hash-frontend/src/pages/block-generator/preview/public/logo512.png
