+++ apps/hash-frontend/src/pages/block-generator/preview/public/manifest.json
