+++ apps/hash-frontend/src/pages/block-generator/preview/src/reportWebVitals.js
