--- apps/hash-frontend/src/pages/react-app-generator/preview/App.js
@@ -1,12 +1,0 @@
-import React from "react";
-import Button from "@mui/material/Button";
-
-const MyButton = () => {
-  return (
-    <Button variant="contained" color="primary">
-      Click me!a
-    </Button>
-  );
-};
-
-export default MyButton;
